==> 3. Counties.Rmd <==
---
title: "Vodaphone Urban/Rural Mobile Coverage"
author: "Guy Weir"
date: "2023-09-28"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = F, message = F)

#import the data
library(tidyverse)
library(readxl)
library(aws.s3)
library(kableExtra)
library(reactable)
library(downloadthis)


colourlist <- c("#660A0B","#A44200","#d99ec7","#2D2D2A","#2C363F", "#FFA3A3", "#ADB8E6", "#DCD6F7")

ofcom.df <- readRDS("ofcom constituency df.rds")
# ofcom.df <- ofcom.df %>% filter(!is.na(`Urban/Rural GB`))

ofcom.df$`Rural (both)` <- ifelse(ofcom.df$`Urban/Rural GB` %in% c("Predominantly Rural", "Urban with Significant Rural"), "Rural (both)", "Urban" )

#table function1
table.fn <- function(table){
  kable(table, format.args = list(big.mark = ",")) %>% 
  kable_styling(bootstrap_options = c("striped", "hover"), 
                full_width = F, 
                position = "left") %>%
  row_spec(0, background = colourlist[1], color = "white") %>% 
  row_spec(1:nrow(table), color = "black")
  
}

#reactable table function

rtable.fn.short <- function(my.df){  my.df %>% 
  reactable(defaultPageSize = 19,
            highlight = T,
            striped = T,
            theme = reactableTheme(
              stripedColor = "#E5E4E2",
              highlightColor = "#8c8c8c",
              cellPadding = "6px 10px",
              style = list(fontFamily = "Arial", fontSize = "12px"),
              #searchInputStyle = list(width = "100%", fontWeight = "400"),
              headerStyle = list(color = "white",background = colourlist[1],
                                 "&:hover[aria-sort]" = list(background = "#8c8c8c "),
                                 "&[aria-sort='ascending'], &[aria-sort='descending']" = list(background = "#8c8c8c"),
                                 borderColor = "#8c8c8c"
              )
            ), filterable = F, sortable = T, pagination = T) 

}

rtable.fn.long <- function(my.df){  my.df %>% 
  reactable(defaultPageSize = 19,
            highlight = T,
            striped = T,
            theme = reactableTheme(
              stripedColor = "#E5E4E2",
              highlightColor = "#8c8c8c",
              cellPadding = "6px 10px",
              style = list(fontFamily = "Arial", fontSize = "12px"),
              #searchInputStyle = list(width = "100%", fontWeight = "400"),
              headerStyle = list(color = "white",background = colourlist[1],
                                 "&:hover[aria-sort]" = list(background = "#8c8c8c "),
                                 "&[aria-sort='ascending'], &[aria-sort='descending']" = list(background = "#8c8c8c"),
                                 borderColor = "#8c8c8c"
              )
            ), filterable = T, sortable = T, pagination = F, height = 600) 

}



```

<!-- ### Urban Rural 4G % outside premises -->

<!-- ```{r UR4G, warning=FALSE} -->
<!-- T_4G_UR <- ofcom.df %>%  -->
<!--   group_by(`Urban/Rural GB`) %>%  -->
<!--   summarise("Not-spots (%)" = round(sum(`4G_prem_out_None_count`)/sum(prem_count)*100,1), -->
<!--             "Partial not-spots (%)" =  round(sum(`4G_prem_1to3_count`)/sum(prem_count)*100,1), -->
<!--             "All operators (%)" =  round(sum(`4G_prem_out_All_count`)/sum(prem_count)*100,1),  -->
<!--             #"Premises count" = sum(prem_count), -->
<!--             "Population 2021" = sum(`Total population 2021`)) %>% na.omit(`Rural (both)`)  -->

<!-- #  -->
<!-- # table.fn(T_4G_UR) -->

<!-- rtable.fn.short(T_4G_UR) -->
<!--   downloadthis::download_this(T_4G_UR, output_extension = ".xlsx")  -->
<!-- # reactable::reactable(T_4G_UR) -->

<!-- ``` -->

<!-- ### Urban Rural 5G % outside premises -->

<!-- ```{r UR5G, warning=FALSE} -->
<!-- T_5G_UR <- ofcom.df %>%  -->
<!--   group_by(`Urban/Rural GB`) %>%  -->
<!--   summarise("Not-spots (%)" = round(sum(`5G_high_confidence_prem_out_None_count`)/sum(prem_count) *100,1), -->
<!--             "Partial not-spots (%)" =  round(sum(`5G_high_confidence_prem_1to3_count`)/sum(prem_count) *100,1), -->
<!--             "All operators (%)" =  round(sum(`5G_high_confidence_prem_out_All_count`)/sum(prem_count) *100,1),  -->
<!--             #"Premises count" = sum(prem_count), -->
<!--             "Population 2021" = sum(`Total population 2021`)) %>% na.omit(`Rural (both)`)  -->

<!-- rtable.fn.short(T_5G_UR) -->
<!-- downloadthis::download_this(T_5G_UR, output_extension = ".xlsx") -->

<!-- ``` -->

<!-- ### Deprivation grouped 40:20:40 and Urban Rural 4G, %ge of outside premises -->
<!-- ```{r deprivation4g402040, warning=FALSE} -->

<!-- T_4G_URD <- ofcom.df %>%  -->
<!--   group_by(`Urban/Rural GB`, `Deprivation group`) %>%  -->
<!--   summarise("Not-spots (%)" = round(sum(`4G_prem_out_None_count`)/sum(prem_count)*100,1), -->
<!--             "Partial not-spots (%)" =  round(sum(`4G_prem_1to3_count`)/sum(prem_count)*100,1), -->
<!--             "All operators (%)" =  round(sum(`4G_prem_out_All_count`)/sum(prem_count)*100,1),  -->
<!--             #"Premises count" = sum(prem_count), -->
<!--             "Population 2021" = sum(`Total population 2021`)) %>% na.omit(`Rural (both)`)  -->

<!-- rtable.fn.short(T_4G_URD) -->
<!-- downloadthis::download_this(T_4G_URD, output_extension = ".xlsx") -->

<!-- ``` -->


<!-- ### Deprivation and Urban Rural 4G, %ge of outside premises -->
<!-- ```{r deprivation4g, warning=FALSE} -->
<!-- T_4G_URD <- ofcom.df %>%  -->
<!--   group_by(`Urban/Rural GB`,  `IMD quintile within nation`) %>%  -->
<!--   summarise("Not-spots (%)" = round(sum(`4G_prem_out_None_count`)/sum(prem_count)*100,1), -->
<!--             "Partial not-spots (%)" =  round(sum(`4G_prem_1to3_count`)/sum(prem_count)*100,1), -->
<!--             "All operators (%)" =  round(sum(`4G_prem_out_All_count`)/sum(prem_count)*100,1),  -->
<!--             #"Premises count" = sum(prem_count), -->
<!--             "Population 2021" = sum(`Total population 2021`)) %>% na.omit(`Rural (both)`)  -->

<!-- rtable.fn.short(T_4G_URD) -->
<!-- downloadthis::download_this(T_4G_URD, output_extension = ".xlsx") -->

<!-- ``` -->

<!-- ### Deprivation 40:20:40 and Urban Rural 5G, %ge of outside premises -->
<!-- ```{r deprivation5g402040, warning=FALSE} -->
<!-- T_5G_URD <- ofcom.df %>%  -->
<!--   group_by(`Urban/Rural GB`, `Deprivation group`) %>%  -->
<!--   summarise("Not-spots (%)" = round(sum(`5G_high_confidence_prem_out_None_count`)/sum(prem_count)*100,1), -->
<!--             "Partial not-spots (%)" =  round(sum(`5G_high_confidence_prem_1to3_count`)/sum(prem_count)*100,1), -->
<!--             "All operators (%)" =  round(sum(`5G_high_confidence_prem_out_All_count`)/sum(prem_count)*100,1),  -->
<!--             #"Premises count" = sum(prem_count), -->
<!--             "Population 2021" = sum(`Total population 2021`)) %>% na.omit(`Rural (both)`)  -->

<!-- rtable.fn.short(T_5G_URD) -->
<!-- downloadthis::download_this(T_5G_URD, output_extension = ".xlsx") -->

<!-- ``` -->


<!-- ### Deprivation and Urban Rural 5G, %ge of outside premises -->
<!-- ```{r deprivation5g, warning=FALSE} -->
<!-- T_5G_URD <- ofcom.df %>%  -->
<!--   group_by(`Urban/Rural GB`, `IMD quintile within nation`) %>%  -->
<!--   summarise("Not-spots (%)" = round(sum(`5G_high_confidence_prem_out_None_count`)/sum(prem_count)*100,1), -->
<!--             "Partial not-spots (%)" =  round(sum(`5G_high_confidence_prem_1to3_count`)/sum(prem_count)*100,1), -->
<!--             "All operators (%)" =  round(sum(`5G_high_confidence_prem_out_All_count`)/sum(prem_count)*100,1),  -->
<!--             #"Premises count" = sum(prem_count), -->
<!--             "Population 2021" = sum(`Total population 2021`)) %>% na.omit(`Rural (both)`)  -->

<!-- rtable.fn.short(T_5G_URD) -->
<!-- downloadthis::download_this(T_5G_URD, output_extension = ".xlsx") -->

<!-- ``` -->

<!-- ### All constituency summary data  -->
<!-- ```{r pcon, warning=FALSE} -->
<!-- pcon.table <- ofcom.df %>% ungroup() %>%   -->
<!--   mutate("Poor coverage percentile (1 is poor)" = ntile(`PC1`,100), -->
<!--          "Poor coverage x Deprivation percentile (1 is poor)" = ntile(index_score,100)) %>% arrange(index_score) %>%  -->

<!--   select("Constituency" = parl_const_name, -->
<!--                                   "Rural/Urban group" = `Urban/Rural GB`, -->
<!--                                   "Coverage / Deprivation index percentile (1 is worst)" =`Poor coverage x Deprivation percentile (1 is poor)`, -->
<!--                                   "Poor coverage percentile (1 is worst)"= `Poor coverage percentile (1 is poor)`, -->
<!--                                   "Population 2021" = `Total population 2021`, -->
<!--                                   "Deprivation quintile 1 is lowest" = `IMD quintile within nation`, -->

<!--                                   "4G Not-spots (%)" =`4G_prem_out_None`, -->
<!--                                   "4G partial not-spots (%)" =`4G_prem_1to3`, -->
<!--                                   "4G All operators (%)" =`4G_prem_out_All`, -->
<!--                                   "5G Not-spots (%)" = `5G_high_confidence_prem_out_None`, -->
<!--                                   "5G partial not-spots (%)" = `5G_high_confidence_prem_1to3`, -->
<!--                                   "5G All operators (%)" = `5G_high_confidence_prem_out_All` -->



<!--          ) -->

<!-- rtable.fn.long(pcon.table) -->
<!-- downloadthis::download_this(pcon.table, output_extension = ".xlsx") -->
<!-- ``` -->


### Counties % premises counted as 5G not-spots by Urban/Rural
Note constituencies do not map into Counties perfectly, and I have allocated counties to constituencies based on the highest count of postcodes.
```{r cnty, warning=FALSE}



county.table <- ofcom.df %>% ungroup() %>%  
  mutate("Poor coverage percentile (1 is poor)" = ntile(`PC1`,100),
         "Poor coverage x Deprivation percentile (1 is poor)" = ntile(index_score,100),

  
         "coverge30" = case_when(`Poor coverage percentile (1 is poor)` <= 30 ~ "Poor coverage",
                        `Poor coverage percentile (1 is poor)` > 30 ~ "Not poor coverage"),
         county_areas = case_when(
          str_detect(UTLA21NM, "Cumbria") ~ "Cumbria",
          str_detect(UTLA21NM,"Suffolk") ~ "Suffolk",
          str_detect(UTLA21NM,"Norfolk") ~ "Norfolk",
          str_detect(UTLA21NM,"Somerset") ~ "Somerset",
          str_detect(UTLA21NM,"Cornwall") ~ "Cornwall",
          str_detect(UTLA21NM,"Devon") ~ "Devon",
          str_detect(Country,"S92000003") & `Rural (both)` == "Rural (both)" ~ "Scotland (rural)",
          str_detect(Country, "W92000004") & `Rural (both)` == "Rural (both)" ~ "Wales (rural)",
          TRUE ~ NA
         )) %>% 
           
           group_by(county_areas, `Rural (both)`) %>% 
           
           summarise(
            "Not-spots (%)" = round(sum(`5G_high_confidence_prem_out_None_count`)/sum(prem_count)*100,1),
             
           ) %>% ungroup() %>% na.omit(`Rural (both)`) %>% pivot_wider(id_cols = county_areas, 
                             values_from = `Not-spots (%)`,
                             names_from = `Rural (both)`) 
  
  




rtable.fn.long(county.table)
downloadthis::download_this(county.table, output_extension = ".xlsx")
```


### Cumbria summary data 
```{r Cumbria, warning=FALSE}
T_cumbria <- ofcom.df %>% ungroup() %>%  
  mutate("Poor coverage percentile (1 is poor)" = ntile(`PC1`,100),
         "Poor coverage x Deprivation percentile (1 is poor)" = ntile(index_score,100)) %>% arrange(index_score) %>% 

  filter(str_detect(UTLA21NM,"Cumbria")) %>% 
  
  select("Constituency" = parl_const_name,
                                  "Rural/Urban group" = `Urban/Rural GB`,
                                  "Coverage / Deprivation index percentile (1 is worst)" =`Poor coverage x Deprivation percentile (1 is poor)`,
                                  "Poor coverage percentile (1 is worst)"= `Poor coverage percentile (1 is poor)`,
                                 
                                  "Deprivation quintile 1 is lowest" = `IMD quintile within nation`,
         
                                 "5G Not-spots (%)" = `5G_high_confidence_prem_out_None`,
                                  "5G partial not-spots (%)" = `5G_high_confidence_prem_1to3`)
                                 
  


rtable.fn.long(T_cumbria)
downloadthis::download_this(T_cumbria, output_extension = ".xlsx")
```


### Suffolk summary data 
```{r Suffolk, warning=FALSE}
T_Suffolk <- ofcom.df %>% ungroup() %>%  
  mutate("Poor coverage percentile (1 is poor)" = ntile(`PC1`,100),
         "Poor coverage x Deprivation percentile (1 is poor)" = ntile(index_score,100)) %>% arrange(index_score) %>% 

  filter(str_detect(UTLA21NM,"Suffolk")) %>% 
  
  select("Constituency" = parl_const_name,
                                  "Rural/Urban group" = `Urban/Rural GB`,
                                  "Coverage / Deprivation index percentile (1 is worst)" =`Poor coverage x Deprivation percentile (1 is poor)`,
                                  "Poor coverage percentile (1 is worst)"= `Poor coverage percentile (1 is poor)`,
                                 
                                  "Deprivation quintile 1 is lowest" = `IMD quintile within nation`,
         
                                 "5G Not-spots (%)" = `5G_high_confidence_prem_out_None`,
                                  "5G partial not-spots (%)" = `5G_high_confidence_prem_1to3`)
                                 
  


rtable.fn.long(T_Suffolk)
downloadthis::download_this(T_Suffolk, output_extension = ".xlsx")
```

### Norfolk summary data 
```{r Norfolk, warning=FALSE}
T_norfolk <- ofcom.df %>% ungroup() %>%  
  mutate("Poor coverage percentile (1 is poor)" = ntile(`PC1`,100),
         "Poor coverage x Deprivation percentile (1 is poor)" = ntile(index_score,100)) %>% arrange(index_score) %>% 

  filter(str_detect(UTLA21NM,"Norfolk")) %>% 
  
  select("Constituency" = parl_const_name,
                                  "Rural/Urban group" = `Urban/Rural GB`,
                                  "Coverage / Deprivation index percentile (1 is worst)" =`Poor coverage x Deprivation percentile (1 is poor)`,
                                  "Poor coverage percentile (1 is worst)"= `Poor coverage percentile (1 is poor)`,
                                 
                                  "Deprivation quintile 1 is lowest" = `IMD quintile within nation`,
         
                                 "5G Not-spots (%)" = `5G_high_confidence_prem_out_None`,
                                  "5G partial not-spots (%)" = `5G_high_confidence_prem_1to3`)
                                 
  


rtable.fn.long(T_norfolk)
downloadthis::download_this(T_norfolk, output_extension = ".xlsx")
```


### Somerset summary data 
```{r Somerset, warning=FALSE}
T_Somerset <- ofcom.df %>% ungroup() %>%  
  mutate("Poor coverage percentile (1 is poor)" = ntile(`PC1`,100),
         "Poor coverage x Deprivation percentile (1 is poor)" = ntile(index_score,100)) %>% arrange(index_score) %>% 

  filter(str_detect(UTLA21NM,"Somerset")) %>% 
  
  select("Constituency" = parl_const_name,
                                  "Rural/Urban group" = `Urban/Rural GB`,
                                  "Coverage / Deprivation index percentile (1 is worst)" =`Poor coverage x Deprivation percentile (1 is poor)`,
                                  "Poor coverage percentile (1 is worst)"= `Poor coverage percentile (1 is poor)`,
                                 
                                  "Deprivation quintile 1 is lowest" = `IMD quintile within nation`,
         
                                 "5G Not-spots (%)" = `5G_high_confidence_prem_out_None`,
                                  "5G partial not-spots (%)" = `5G_high_confidence_prem_1to3`)
                                 
  


rtable.fn.long(T_Somerset)
downloadthis::download_this(T_Somerset, output_extension = ".xlsx")
```


### Cornwall summary data 
```{r Cornwall, warning=FALSE}
Cornwall <- ofcom.df %>% ungroup() %>%  
  mutate("Poor coverage percentile (1 is poor)" = ntile(`PC1`,100),
         "Poor coverage x Deprivation percentile (1 is poor)" = ntile(index_score,100)) %>% arrange(index_score) %>% 

  filter(str_detect(UTLA21NM,"Cornwall")) %>% 
  
  select("Constituency" = parl_const_name,
                                  "Rural/Urban group" = `Urban/Rural GB`,
                                  "Coverage / Deprivation index percentile (1 is worst)" =`Poor coverage x Deprivation percentile (1 is poor)`,
                                  "Poor coverage percentile (1 is worst)"= `Poor coverage percentile (1 is poor)`,
                                 
                                  "Deprivation quintile 1 is lowest" = `IMD quintile within nation`,
         
                                 "5G Not-spots (%)" = `5G_high_confidence_prem_out_None`,
                                  "5G partial not-spots (%)" = `5G_high_confidence_prem_1to3`)
                                 
  


rtable.fn.long(Cornwall)
downloadthis::download_this(Cornwall, output_extension = ".xlsx")
```

### Devon summary data 
```{r Devon, warning=FALSE}
Devon <- ofcom.df %>% ungroup() %>%  
  mutate("Poor coverage percentile (1 is poor)" = ntile(`PC1`,100),
         "Poor coverage x Deprivation percentile (1 is poor)" = ntile(index_score,100)) %>% arrange(index_score) %>% 

  filter(str_detect(UTLA21NM,"Devon")) %>% 
  
  select("Constituency" = parl_const_name,
                                  "Rural/Urban group" = `Urban/Rural GB`,
                                  "Coverage / Deprivation index percentile (1 is worst)" =`Poor coverage x Deprivation percentile (1 is poor)`,
                                  "Poor coverage percentile (1 is worst)"= `Poor coverage percentile (1 is poor)`,
                                 
                                  "Deprivation quintile 1 is lowest" = `IMD quintile within nation`,
         
                                 "5G Not-spots (%)" = `5G_high_confidence_prem_out_None`,
                                  "5G partial not-spots (%)" = `5G_high_confidence_prem_1to3`)
                                 
  


rtable.fn.long(Devon)
downloadthis::download_this(Devon, output_extension = ".xlsx")
```

### Scotland summary data 
```{r Scotland, warning=FALSE}
Scotland <- ofcom.df %>% ungroup() %>%  
  mutate("Poor coverage percentile (1 is poor)" = ntile(`PC1`,100),
         "Poor coverage x Deprivation percentile (1 is poor)" = ntile(index_score,100)) %>% arrange(index_score) %>% 

  filter(is.na(UTLA21NM) & `Rural (both)` == "Rural (both)") %>% 
  
  select("Constituency" = parl_const_name,
                                  "Rural/Urban group" = `Urban/Rural GB`,
                                  "Coverage / Deprivation index percentile (1 is worst)" =`Poor coverage x Deprivation percentile (1 is poor)`,
                                  "Poor coverage percentile (1 is worst)"= `Poor coverage percentile (1 is poor)`,
                                 
                                  "Deprivation quintile 1 is lowest" = `IMD quintile within nation`,
         
                                 "5G Not-spots (%)" = `5G_high_confidence_prem_out_None`,
                                  "5G partial not-spots (%)" = `5G_high_confidence_prem_1to3`)
                                 
  


rtable.fn.long(Scotland)
downloadthis::download_this(Scotland, output_extension = ".xlsx")
```

### Wales summary data 
```{r Wales, warning=FALSE}
Wales <- ofcom.df %>% ungroup() %>%  
  mutate("Poor coverage percentile (1 is poor)" = ntile(`PC1`,100),
         "Poor coverage x Deprivation percentile (1 is poor)" = ntile(index_score,100)) %>% arrange(index_score) %>% 

  filter(Country == "W92000004" & `Rural (both)` == "Rural (both)") %>% 
  
  select("Constituency" = parl_const_name,
                                  "Rural/Urban group" = `Urban/Rural GB`,
                                  "Coverage / Deprivation index percentile (1 is worst)" =`Poor coverage x Deprivation percentile (1 is poor)`,
                                  "Poor coverage percentile (1 is worst)"= `Poor coverage percentile (1 is poor)`,
                                 
                                  "Deprivation quintile 1 is lowest" = `IMD quintile within nation`,
         
                                 "5G Not-spots (%)" = `5G_high_confidence_prem_out_None`,
                                  "5G partial not-spots (%)" = `5G_high_confidence_prem_1to3`)
                                 
  


rtable.fn.long(Wales)
downloadthis::download_this(Wales, output_extension = ".xlsx")
```
